backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic
openpyxl
python-multipart
